+++ packages/components/package.json
@@ -6,7 +6,7 @@
   "module": "dist/index.js",
   "dependencies": {
     "@penrose/core": "^1.3.0",
-    "styled-components": "^5.2.1"
+    "styled-components": "^4.4.1"
   },
   "peerDependencies": {
     "react": "^16.13.1",
