+++ packages/components/src/Embed.tsx
@@ -1,8 +1,34 @@
 import "./Embed.css";
-import React from "react";
+import React, { CSSProperties } from "react";
 import Logo from "./icons/Logo";
 import Resample from "./icons/Resample";
 import { ISimpleProps, Simple } from "./Simple";
+
+const footerStyle: CSSProperties = {
+  display: "flex",
+  overflow: "hidden",
+  justifyContent: "space-between",
+  padding: "0 1em",
+  alignItems: "center",
+  backgroundColor: "#40b4f7",
+  borderRadius: "0 0 9px 9px",
+};
+
+const containerStyle: CSSProperties = {
+  position: "relative",
+  borderRadius: "10px",
+  border: "0.5px solid rgba(0, 0, 0, 0.2)",
+  boxShadow: "0 5px 8px 0 rgba(0, 0, 0, 0.2)",
+  overflow: "hidden",
+};
+
+const textStyle = {
+  fontFamily: '"Open Sans", sans-serif',
+  color: "white",
+  margin: "0.7em",
+  // fontWeight: "lighter",
+  fontSize: "smaller",
+};
 
 const Heart = () => (
   <span role="img" aria-label="heart">
@@ -20,14 +46,14 @@ class Embed extends React.Component<ISimpleProps> {
 
   render = () => {
     return (
-      <div className="embed-container">
+      <div className="embed-container" style={containerStyle}>
         <Simple {...this.props} ref={this.simpleRef} />
         <div style={{ width: "100%", height: "100%" }} />
-        <div className="embed-footer">
+        <div className="embed-footer" style={footerStyle}>
           <a href="https://penrose.ink">
             <Logo width={24} color={"white"} />
           </a>
-          <div className="embed-text">
+          <div className="embed-text" style={textStyle}>
             Generated by Penrose with <Heart />
           </div>
           <div onClick={this.resample}>
